% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_norm.R
\name{load_normalize}
\alias{load_normalize}
\title{Load normalize proteomic data}
\usage{
load_normalize(proteins_raw, peptides_raw, num_input, num_group, num_exp,
  num_pool)
}
\arguments{
\item{proteins_raw}{A raw data frame with protein abundances}

\item{peptides_raw}{A raw data frame with peptide abundances}

\item{num_input}{Number of input kits in the experiment}

\item{num_group}{Number of experimental plexes}

\item{num_exp}{Number of experimental subgroups per plex}

\item{num_pool}{Number of num_pool groups}
}
\value{
data_load_normalized return the load normalized proteomic data
}
\description{
Load normalizes searched Proteome Discoverer data
}
