% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_pca.R
\name{protein_pca}
\alias{protein_pca}
\title{Principle component analysis}
\usage{
protein_pca(tidyData, components = 4, center = FALSE, scale = FALSE)
}
\arguments{
\item{tidyData}{Normalized data to perform the PCA analysis on}

\item{components}{The number of components}

\item{center}{Whether the data is centered or not}

\item{scale}{Whether the data is scaled or not}
}
\value{
A list containing the principal components and the ggplots of the score plot and the scree plot
}
\description{
: Returns an exploratory PCA analysis, along with a score plot and a scree plot
}
\author{
Matthew Huang
}
